权限

+++ src/main/java/com/github/microprograms/qipai_exchange_manager_api/public_api/ChoiceBanner_UpdateAll_Api.java
@@ -8,11 +8,13 @@ import com.github.microprograms.ignite_utils.sql.dml.Condition;
 import com.github.microprograms.ignite_utils.sql.dml.DeleteSql;
 import com.github.microprograms.ignite_utils.sql.dml.InsertSql;
 import com.github.microprograms.micro_api_runtime.annotation.MicroApiAnnotation;
+import com.github.microprograms.micro_api_runtime.exception.MicroApiExecuteException;
 import com.github.microprograms.micro_api_runtime.model.Request;
 import com.github.microprograms.micro_api_runtime.model.Response;
 import com.github.microprograms.micro_entity_definition_runtime.annotation.Comment;
 import com.github.microprograms.micro_entity_definition_runtime.annotation.Required;
 import com.github.microprograms.qipai_exchange_manager_api.public_api.Banner_QueryAll_Api.Resp;
+import com.github.microprograms.qipai_exchange_manager_api.utils.Commons;
 import com.github.microprograms.qipai_exchange_manager_api.utils.Consts;
 
 @Comment(value = "会员专区Banner - 更新全部")
@@ -22,6 +24,17 @@ public class ChoiceBanner_UpdateAll_Api {
     public static Response execute(Request request) throws Exception {
         Req req = (Req) request;
         Resp resp = new Resp();
+        if (StringUtils.isBlank(req.getToken())) {
+            throw new MicroApiExecuteException(ErrorCodeEnum.missing_required_parameters);
+        }
+        DepartmentMember departmentMember = Commons.queryDepartmentMemberByToken(req.getToken());
+        if (departmentMember == null) {
+            throw new MicroApiExecuteException(ErrorCodeEnum.invalid_token);
+        }
+        Department department = Commons.queryDepartmentById(departmentMember.getDepartmentId());
+        if (!Commons.hasPermission(department, PermissionEnum.choiceBannerManage)) {
+            throw new MicroApiExecuteException(ErrorCodeEnum.permission_denied);
+        }
         try (Connection conn = IgniteUtils.getConnection(Consts.jdbc_url)) {
             conn.createStatement().executeUpdate(new DeleteSql(Banner.class).where(Condition.build("type=", 2)).build());
             for (Banner x : req.getBanners()) {
